+++ MaiMai/MaiMai/Content/resource_nico/createProductPost.css
@@ -1,0 +1,4 @@
+.formFrame{
+    width: 100%;
+
+}
