+++ MaiMai/MaiMai/packages.config
@@ -2,9 +2,10 @@
 <packages>
   <package id="Antlr" version="3.5.0.2" targetFramework="net472" />
   <package id="bootstrap" version="4.6.0" targetFramework="net472" />
+  <package id="DocumentFormat.OpenXml" version="2.12.2" targetFramework="net472" />
   <package id="EntityFramework" version="6.2.0" targetFramework="net472" />
   <package id="EntityFramework.zh-Hant" version="6.2.0" targetFramework="net472" />
-  <package id="jQuery" version="3.4.1" targetFramework="net472" />
+  <package id="jQuery" version="3.4.0" targetFramework="net472" />
   <package id="jQuery.UI.Combined" version="1.12.1" targetFramework="net472" />
   <package id="jQuery.Validation" version="1.17.0" targetFramework="net472" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.7" targetFramework="net472" />
